requirements.txt
gradio
tensorflow-cpu==2.15.0
opencv-python-headless
numpy
